322、商城业务-秒杀服务-秒杀流程

+++ mall-seckill/src/main/kotlin/com/vurtnewk/mall/seckill/to/SecKillSkuRedisDto.kt
@@ -33,7 +33,7 @@ data class SecKillSkuRedisDto(
     /**
      * 每人限购数量
      */
-    var seckillLimit: BigDecimal? = null,
+    var seckillLimit: Int? = null,
     /**
      * 排序
      */
